+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 # 2.3.0 (unreleased)
 
 * Drop support for PHP 8.1
+* Add support for psalm/phpstan types in docblocks (e.g. `list<string>` or `array<string, int>`)
 
 # 2.2.1